+++ R/imports.R
@@ -1,5 +1,0 @@
-.onUnload <- function(libpath) library.dynam.unload("adfExplorer2", libpath)
-
-#' @useDynLib adfExplorer2, .registration = TRUE
-#' @importFrom methods setOldClass
-NULL