+++ NEWS.md
@@ -1,7 +1,9 @@
-adfExplorer2 v0.0.2.0001
+adfExplorer2 v0.0.3
 -------------
 
   * Cleaned out redundant code
+  * Some debugging with valgrind and lldb
+  * User can now rename a entries
 
 adfExplorer2 v0.0.2
 -------------